Finish login-modal to login refactor & optimise imports
